Update and rename 2018_02_07_110212_drop_content_template_table.php to 2018_02_07_112015_drop_content_template_table

+++ database/migrations/2018_02_07_112015_drop_content_template_table
@@ -13,11 +13,18 @@ class DropContentTemplateTable extends Migration
      */
     public function up()
     {
-        Schema::table('courses', function (Blueprint $table) {
-            $table->dropForeign('template_ibfk_1');
-        });
 
-        Schema::dropIfExists('content_templates');
+        Schema::table('courses', function (Blueprint $table) {
+            $table->dropIndex('template_ibfk_1');
+            $table->dropForeign('template_ibfk_1');
+
+            $table->dropcolumn('template_id');
+        });
+        
+        if(Schema::hasTable('content_templates')){
+            Schema::dropIfExists('content_templates');
+        }
+        
     }
 
     /**
